requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
psycopg2-binary
faker
logging